rename post_sync.connector -> post_sync.connectors (per spec)
