+++ user_sync/post_sync/connectors/__init__.py
@@ -1,4 +1,4 @@
-from .sign import connector_sign
+from .sign_sync import connector_sign
 
 __CONNECTORS__ = {
     'sign_sync': connector_sign.SignConnector
